🧪 One Token Model - Pallet Funding Tests (#402)
## What?
- Change mock AccountId from u32 to u64
- Test the OTM flow inside pallet-funding

## Why?
- PalletId account derivations start with a 32bit constant, so that meant the proxy-bonding escrow and the project_id escrow had the same account
- So far we only tested the proxy-bonding in a vacuum.

## Testing?
- `one_token_mode_bid_funding_success`
- `one_token_mode_bid_funding_failed`
- `one_token_mode_contribution_funding_success`
- `one_token_mode_contribution_funding_failed`

## Anything Else?
e2e tests need to be updated with OTM values. In another PR

+++ pallets/funding/src/tests/2_evaluation.rs
@@ -764,7 +764,7 @@ mod evaluate_extrinsic {
 			let mut inst = MockInstantiator::new(Some(RefCell::new(new_test_ext())));
 			let project_metadata = default_project_metadata(ISSUER_1);
 			let evaluations = (0u32..<TestRuntime as Config>::MaxEvaluationsPerProject::get())
-				.map(|i| UserToUSDBalance::<TestRuntime>::new(i as u32 + 420u32, (100u128 * CT_UNIT).into()))
+				.map(|i| UserToUSDBalance::<TestRuntime>::new(i as u64 + 420, (100u128 * CT_UNIT).into()))
 				.collect_vec();
 			let failing_evaluation = UserToUSDBalance::new(EVALUATOR_1, 1000 * CT_UNIT);
